+++ src/cheat/precompiled.h
@@ -254,6 +254,7 @@ namespace hl
 #include "features/FrameSkipper.h"
 #include "features/ESP.h"
 #include "features/VanillaCrosshair.h"
+#include "features/StuffCmdFilter.h"
 
 // ui
 #include "ui/IRendererContext.h"